refactor: extract pytest independent code out into SnapshotSession
