+++ changelog.d/20250914_233146_15r10nk-git_snapshot_session.md
@@ -1,0 +1,3 @@
+### Fixed
+
+- handle in cases where `pytest_unconfigure` is called but not `pytest_configure` ([#293](https://github.com/15r10nk/inline-snapshot/issues/293))
